refactor: separate audio dropdown from controller
